move NodeInfo to domain
